Implementing versioning in Player.java and Achievement.java

+++ src/main/java/org/springframework/samples/SevenIslands/player/PlayerController.java
@@ -34,6 +34,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 
 @Controller
@@ -307,9 +308,10 @@ public class PlayerController {
 
     @GetMapping(path="/edit/{playerId}")
     public String updatePlayer(@PathVariable("playerId") int playerId, ModelMap model) {
+
         Optional<Player> player = playerService.findPlayerById(playerId); // optional puede ser error el import
         String view = VIEWS_PLAYERS_CREATE_OR_UPDATE_FORM;
-       securityService.insertIdUserModelMap(model);
+        securityService.insertIdUserModelMap(model);
         //Test if currentplayer is admin or the same id
         // TODO: Comprobar que sea o admin o q el usuer registrado tenga el mismo id q el de la url
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -355,7 +357,8 @@ public class PlayerController {
      */
 
     @PostMapping(value = "/edit/{playerId}")
-	public String processUpdateForm(@Valid Player player, BindingResult result,@PathVariable("playerId") int playerId, ModelMap model) {
+	public String processUpdateForm(@Valid Player player, BindingResult result,@PathVariable("playerId") int playerId, ModelMap model,
+                                        @RequestParam(value="version", required = false) Integer version) {
 
         
         if(playerService.playerHasInappropiateWords(player)){
@@ -371,44 +374,50 @@ public class PlayerController {
 		}
 		else {
                     
-                    Player playerToUpdate=this.playerService.findPlayerById(playerId).get();
-                    int id = playerToUpdate.getUser().getAuthorities().iterator().next().getId();
-                    String username = playerToUpdate.getUser().getUsername();
-                    Iterable<Player> players = playerService.findAll();
-                    List<String> usernames = StreamSupport.stream(players.spliterator(),false).map(x->x.getUser().getUsername().toString()).collect(Collectors.toList());
-                    //borrar user antes de grabarlo en playerToUpdate
-                    //validador
-			BeanUtils.copyProperties(player, playerToUpdate,"id", "profilePhoto","totalGames","totalTimeGames","avgTimeGames","maxTimeGame","minTimeGame","totalPointsAllGames","avgTotalPoints","favoriteIsland","favoriteTreasure","maxPointsOfGames","minPointsOfGames","achievements","cards","watchGames","forums","games","invitations","friend_requests","players_friends","gamesCreador");  //METER AQUI OTRAS PROPIEDADES                                                                                
-                    try {                    
-                        //this.playerService.savePlayer(playerToUpdate);
-                        
-                        //If the username is already in the DB and it's was edited then it means that
-                        //we are overwritting another user
-                        if(usernames.stream().anyMatch(x->x.equals(playerToUpdate.getUser().getUsername()))&&
-                        !playerToUpdate.getUser().getUsername().equals(username)){
-                            return "errors/error-500";
-                        }
+            Player playerToUpdate=this.playerService.findPlayerById(playerId).get();
+            int id = playerToUpdate.getUser().getAuthorities().iterator().next().getId();
+            String username = playerToUpdate.getUser().getUsername();
+            Iterable<Player> players = playerService.findAll();
+            List<String> usernames = StreamSupport.stream(players.spliterator(),false).map(x->x.getUser().getUsername().toString()).collect(Collectors.toList());
+            //borrar user antes de grabarlo en playerToUpdate
+            //validador
 
-                        this.playerService.savePlayer(playerToUpdate);
-                        authoritiesService.deleteAuthorities(id);
-                        if(username != playerToUpdate.getUser().getUsername()){
-                            userService.delete(username);
-                        }      
+            if(playerToUpdate.getVersion()!=version){    //Version
+                model.put("message", "Concurrent modification of player! Try again!");
+                return updatePlayer(playerToUpdate.getId(),model);
+            }
 
-                    } catch (Exception ex) {
-                        result.rejectValue("name", "duplicate", "already exists");
-                        return VIEWS_PLAYERS_CREATE_OR_UPDATE_FORM;
-                    }
-                    if(securityService.isAdmin()){
-                        return "redirect:/players";
-                    }else{
-                        if(username != playerToUpdate.getUser().getUsername()){
-                            SecurityContextHolder.getContext().getAuthentication().setAuthenticated(false);
-                            return "redirect:/welcome";
-                        }
-                        return "redirect:/players/profile/{playerId}";
-                        
-                    }
+            BeanUtils.copyProperties(player, playerToUpdate,"id", "in_game","statistic","achievements","cards","watchGames","forums","games","gamesCreador");  //METER AQUI OTRAS PROPIEDADES                                                                                
+            try {                    
+                //this.playerService.savePlayer(playerToUpdate);
+                
+                //If the username is already in the DB and it's was edited then it means that
+                //we are overwritting another user
+                if(usernames.stream().anyMatch(x->x.equals(playerToUpdate.getUser().getUsername()))&&
+                !playerToUpdate.getUser().getUsername().equals(username)){
+                    return "errors/error-500";
+                }
+
+                this.playerService.savePlayer(playerToUpdate);
+                authoritiesService.deleteAuthorities(id);
+                if(username != playerToUpdate.getUser().getUsername()){
+                    userService.delete(username);
+                }      
+
+            } catch (Exception ex) {
+                result.rejectValue("name", "duplicate", "already exists");
+                return VIEWS_PLAYERS_CREATE_OR_UPDATE_FORM;
+            }
+            if(securityService.isAdmin()){
+                return "redirect:/players";
+            }else{
+                if(username != playerToUpdate.getUser().getUsername()){
+                    SecurityContextHolder.getContext().getAuthentication().setAuthenticated(false);
+                    return "redirect:/welcome";
+                }
+                return "redirect:/players/profile/{playerId}";
+                
+            }
 		}
 	}  
 }
